+++ IsekaiMod/Content/Classes/IsekaiProtagonist/Archetypes/EdgeLord/EdgeLordArchetype.cs
@@ -1,4 +1,5 @@
-﻿using IsekaiMod.Content.Features.IsekaiProtagonist.Archetypes.EdgeLord;
+﻿using IsekaiMod.Content.Features.IsekaiProtagonist;
+using IsekaiMod.Content.Features.IsekaiProtagonist.Archetypes.EdgeLord;
 using IsekaiMod.Content.Features.IsekaiProtagonist.InheritedClassFeature;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
@@ -40,6 +41,7 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist.Archetypes {
                 bp.IsDivineCaster = true;
                 bp.RemoveFeatures = new LevelEntry[] {
                     Helpers.CreateLevelEntry(1, IsekaiProtagonistProficiencies, LegacySelection.GetClassFeature()),
+                    Helpers.CreateLevelEntry(3, ReleaseEnergy.Get()),
                     Helpers.CreateLevelEntry(5, OverpoweredAbilitySelection),
                     Helpers.CreateLevelEntry(8, IsekaiFastMovement),
                     Helpers.CreateLevelEntry(10, IsekaiAuraSelection),