+++ IsekaiMod/Content/Features/IsekaiProtagonist/ReleaseEnergy.cs
@@ -1,0 +1,31 @@
+﻿using Kingmaker.Blueprints;
+using Kingmaker.Blueprints.Classes;
+using Kingmaker.UnitLogic.FactLogic;
+using TabletopTweaks.Core.Utilities;
+using UnityEngine;
+using static IsekaiMod.Main;
+
+namespace IsekaiMod.Content.Features.IsekaiProtagonist {
+    class ReleaseEnergy {
+        public static BlueprintFeature Feature;
+
+        private static readonly Sprite Icon_AngelfireApostleChannel = BlueprintTools.GetBlueprint<BlueprintFeature>("9d30d6cc7bfcda44aab7505f7ed3f933").m_Icon;
+        public static void Add() {
+            Feature = Helpers.CreateBlueprint<BlueprintFeature>(IsekaiContext, "ReleaseEnergy", bp => {
+                bp.SetName(IsekaiContext, "Release Energy");
+                bp.SetDescription(IsekaiContext, "The Isekai Protagonist is able to channel both positive energy and negative energy."
+                    + "\nSo this what happens when you channel one percent of your power...");
+                bp.m_Icon = Icon_AngelfireApostleChannel;
+                bp.AddComponent<AddFacts>(c => {
+                    c.m_Facts = new BlueprintUnitFactReference[] {
+                        IsekaiChannelPositiveEnergy.Get().ToReference<BlueprintUnitFactReference>(),
+                        IsekaiChannelNegativeEnergy.Get().ToReference<BlueprintUnitFactReference>(),
+                    };
+                });
+            });
+        }
+        public static BlueprintFeature Get() {
+            return Feature;
+        }
+    }
+}